Adjusted folder structure, renamed css file and CareersPage, deleted App.css (empty)
