+++ adopt-me/src/app/App.jsx
@@ -1,15 +1,15 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import HomePage from "./pages/HomePage.jsx";
-import AboutPage from "./pages/AboutPage.jsx";
-import AnimalsPage from "./pages/AnimalsPage.jsx";
-import ContactPage from "./pages/ContactPage.jsx";
-import AdoptionFormPage from "./pages/AdoptionFormPage.jsx";
+import HomePage from "../pages/HomePage.jsx";
+import AboutPage from "../pages/AboutPage.jsx";
+import AnimalsPage from "../pages/AnimalsPage.jsx";
+import ContactPage from "../pages/ContactPage.jsx";
+import AdoptionFormPage from "../pages/AdoptionFormPage.jsx";
 import Careers from "./pages/Careers.jsx";
-import VolunteerPage from "./pages/VolunteerPage.jsx";
-import Stories from "./pages/StoriesPage.jsx";
-import Footer from "./components/Footer.jsx";
-import Navbar from "./components/Navbar.jsx";
-import Header from "./components/Header.jsx";
+import VolunteerPage from "../pages/VolunteerPage.jsx";
+import Stories from "../pages/StoriesPage.jsx";
+import Footer from "../components/layout/Footer.jsx";
+import Navbar from "../components/layout/Navbar.jsx";
+import Header from "../components/layout/Header.jsx";
 import banner from "./assets/images/banner-graphic.jpg";
 import "./App.css";
