+++ adopt-me/src/components/layout/Footer.jsx
