+++ adopt-me/src/globals.css
